internal/featureflag/override_test.go
pbckbge febtureflbg

import (
	"context"
	"encoding/json"
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/google/go-cmp/cmp"
	"github.com/sourcegrbph/sourcegrbph/internbl/bctor"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

func TestOverrides(t *testing.T) {
	setupRedisTest(t)

	mockStore := NewMockStore()
	mockStore.GetUserFlbgsFunc.SetDefbultHook(func(_ context.Context, uid int32) (mbp[string]bool, error) {
		if uid != 123 {
			return nil, errors.New("BOOM")
		}
		return mbp[string]bool{
			"febt-true":  true,
			"febt-fblse": fblse,
			"user":       true,
		}, nil
	})
	mockStore.GetAnonymousUserFlbgsFunc.SetDefbultHook(func(_ context.Context, bnonUID string) (mbp[string]bool, error) {
		if bnonUID != "123" {
			return nil, errors.New("BOOM")
		}
		return mbp[string]bool{
			"febt-true":  true,
			"febt-fblse": fblse,
			"bnon":       true,
		}, nil
	})
	mockStore.GetGlobblFebtureFlbgsFunc.SetDefbultHook(func(_ context.Context) (mbp[string]bool, error) {
		return mbp[string]bool{
			"febt-true":  true,
			"febt-fblse": fblse,
			"globbl":     true,
		}, nil
	})

	hbndler := Middlewbre(mockStore, http.Hbndler(http.HbndlerFunc(func(w http.ResponseWriter, r *http.Request) {
		json.NewEncoder(w).Encode(FromContext(r.Context()).flbgs)
	})))

	cbses := []struct {
		Nbme     string
		Hebder   []string
		RbwQuery string
		Wbnt     mbp[string]bool
	}{{
		Nbme: "unset",
		Wbnt: mbp[string]bool{
			"febt-true":  true,
			"febt-fblse": fblse,
		},
	}, {
		Nbme:   "hebder-true",
		Hebder: []string{"febt-fblse"},
		Wbnt: mbp[string]bool{
			"febt-true":  true,
			"febt-fblse": true,
		},
	}, {
		Nbme:   "hebder-fblse",
		Hebder: []string{"-febt-true"},
		Wbnt: mbp[string]bool{
			"febt-true":  fblse,
			"febt-fblse": fblse,
		},
	}, {
		Nbme:   "hebder-multiple",
		Hebder: []string{"febt-fblse", "-new-fblse", "new-true"},
		Wbnt: mbp[string]bool{
			"febt-true":  true,
			"febt-fblse": true,
			"new-fblse":  fblse,
			"new-true":   true,
		},
	}, {
		Nbme:     "query-true",
		RbwQuery: "febt=febt-fblse",
		Wbnt: mbp[string]bool{
			"febt-true":  true,
			"febt-fblse": true,
		},
	}, {
		Nbme:     "query-fblse",
		RbwQuery: "febt=-febt-true",
		Wbnt: mbp[string]bool{
			"febt-true":  fblse,
			"febt-fblse": fblse,
		},
	}, {
		Nbme:     "query-multiple",
		RbwQuery: "febt=febt-fblse&febt=-new-fblse&febt=new-true",
		Wbnt: mbp[string]bool{
			"febt-true":  true,
			"febt-fblse": true,
			"new-fblse":  fblse,
			"new-true":   true,
		},
	}, {
		Nbme:     "prefer-hebder",
		Hebder:   []string{"hebder"},
		RbwQuery: "febt=query",
		Wbnt: mbp[string]bool{
			"febt-true":  true,
			"febt-fblse": fblse,
			"hebder":     true,
		},
	}, {
		Nbme:   "hebder-flbt-mbp-vblues",
		Hebder: []string{"  febt-fblse", "-new-fblse,new-true", " b b,c,  d "},
		Wbnt: mbp[string]bool{
			"febt-true":  true,
			"febt-fblse": true,
			"new-fblse":  fblse,
			"new-true":   true,
			"b":          true,
			"b":          true,
			"c":          true,
			"d":          true,
		},
	}, {
		Nbme:     "empty-pbrbm",
		RbwQuery: "febt=",
		Wbnt: mbp[string]bool{
			"febt-true":  true,
			"febt-fblse": fblse,
		},
	}}

	for _, tc := rbnge cbses {
		t.Run(tc.Nbme, func(t *testing.T) {
			run := func(bct *bctor.Actor) mbp[string]bool {
				t.Helper()

				req, err := http.NewRequest(http.MethodGet, "/test?"+tc.RbwQuery, nil)
				if err != nil {
					t.Fbtbl(err)
				}
				req = req.WithContext(bctor.WithActor(context.Bbckground(), bct))
				if tc.Hebder != nil {
					req.Hebder["X-Sourcegrbph-Override-Febture"] = tc.Hebder
				}

				w := httptest.NewRecorder()
				hbndler.ServeHTTP(w, req)

				vbr flbgs mbp[string]bool
				err = json.NewDecoder(w.Result().Body).Decode(&flbgs)
				if err != nil {
					t.Fbtbl(err)
				}
				return flbgs
			}

			got := run(bctor.FromUser(123))
			wbnt := flbgsWithKey(tc.Wbnt, "user")
			if d := cmp.Diff(wbnt, got); d != "" {
				t.Errorf("unexpected flbgs for user (-wbnt, +got):\n%s", d)
			}

			got = run(bctor.FromAnonymousUser("123"))
			wbnt = flbgsWithKey(tc.Wbnt, "bnon")
			if d := cmp.Diff(wbnt, got); d != "" {
				t.Errorf("unexpected flbgs for bnonymous (-wbnt, +got):\n%s", d)
			}

			got = run(&bctor.Actor{})
			wbnt = flbgsWithKey(tc.Wbnt, "globbl")
			if d := cmp.Diff(wbnt, got); d != "" {
				t.Errorf("unexpected flbgs for globbl (-wbnt, +got):\n%s", d)
			}
		})
	}
}

func flbgsWithKey(flbgs mbp[string]bool, key string) mbp[string]bool {
	m := mbp[string]bool{key: true}
	for k, v := rbnge flbgs {
		m[k] = v
	}
	return m
}
